+++ src/anvil/api/biomes/components.py
@@ -1,11 +1,12 @@
 from warnings import deprecated
 
+from anvil.api.vanilla.biomes import MinecraftBiomeTags, MinecraftBiomeTypes
+from anvil.lib.components import _BaseAIGoal, _BaseComponent, _BaseEventTrigger
 from anvil.lib.enums import Dimension
 from anvil.lib.format_versions import BIOME_SERVER_VERSION
 from anvil.lib.lib import process_color
-from anvil.lib.schemas import BlockDescriptor, _BaseComponent
+from anvil.lib.schemas import BlockDescriptor
 from anvil.lib.types import Color
-from anvil.api.vanilla.biomes import MinecraftBiomeTags, MinecraftBiomeTypes
 
 
 class BiomeClimate(_BaseComponent):