+++ src/anvil/lib/blockbench.py
@@ -1,13 +1,8 @@
 import base64
 import json
 import os
-import pprint
-import uuid
-from calendar import c
 from collections import defaultdict
-from copy import deepcopy
-from csv import Error
-from typing import Dict, List, Literal, Tuple, Union
+from typing import Dict, List, Union
 from warnings import warn
 
 from anvil import CONFIG